Use a config file instead of predefining symbols
We often have trouble with .cproject merge conflicts. This will help reduce that
The config file is set as a pre-include, so it's included in all files, no #include directive needed
The predefined symbols were also not self-explanatory, needing a spreadsheet to keep track of them. That spreadsheet is now irrelevant
Configurations can now be easily changed in config.h